Move DateTimeFingerprint to its own file
